% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_countries.R
\name{area_weighted_synthetic_data}
\alias{area_weighted_synthetic_data}
\title{Calculate area weighted synthetic static panel data}
\usage{
area_weighted_synthetic_data(df, static_year, ...)
}
\arguments{
\item{df}{A dynamic panel-data with gwcode and year plus any additional variables.}

\item{static_year}{An integer of the year you want to create a static world template based on.}

\item{...}{Additional parameters, please not use yet as they might not work properly.}
}
\value{
A static panel-data data.frame
}
\description{
This function creates a static-world panel data.frame from dynamic gwcode-year panel-data by weighing
the data from the ancestors of countries that existed at the end of `static_year` using the shares of
ancestor's territories that can be traced to the current countries.
}
\examples{
vdem_dynamic <- get_vdem(v2x_libdem, v2x_regime)
vdem_static <- area_weighted_synthetic_data(vdem_dynamic, 2019)

}
