% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{battlelocationsfatalityshares}
\alias{battlelocationsfatalityshares}
\title{Battle-locations fatality shares data}
\format{
## `battlelocationsfatalityshares`
A data frame with 134 rows and 22 columns:
\describe{
  \item{conflict_id}{UCDP/PRIO conflict id}
  \item{old_id}{Old conflict id as used in PRIO Battle-deaths data}
  \item{year}{Year}
  \item{battle_loc}{cShapes 2.0 gwcodes of countries with any battles within a conflict_id}
  \item{side_a}{Actor on side A}
  \item{side_a_2nd}{States that enter conflict with troops in support of Side A}
  \item{side_b}{Actor on side B}
  \item{side_b_2nd}{States that enter conflict with troops in support of Side B}
  \item{intensity_level}{1: Between 25 and 999 battle-deaths in a given year in the conflict, 2: > 1000 BRDs}
  \item{cumulative_intensity}{Whether conflict since onset has exceeded 1000 BRDs}
  \item{start_date}{Date of first battle-death in conflict}
  \item{ep_end_date}{Date when conflict activity ended}
  \item{bdeadlow}{Battle-deaths, low estimate from PRIO Battle-deaths}
  \item{bdeadhig}{Battle-deaths, high estimate from PRIO Battle-deaths}
  \item{bdeadbes}{Battle-deaths, best estimate from PRIO Battle-deaths}
  \item{territory_name}{The name of the territory over which the conflict is fought}
  \item{gwnoloc}{Country codes of the incompatibility}
  \item{share_a, share_b, share_c, share_d}{The coded share of battle-deaths occuring in 1st, 2nd, 3rd, and 4th gwcode in the battle_loc variable}
  \item{note}{My notes for the coding. These are not replicated for each year if the coding relates to the whole conflict_id.}
}
}
\source{
<https://www.prio.org/data/1>

<https://ucdp.uu.se/downloads/index.html#armedconflict>
}
\usage{
battlelocationsfatalityshares
}
\description{
Manually coded yearly fatality shares for conflicts in UCDP/PRIO ACD version 23.1
with more than one battle-location as coded in \code{\link[=ucdp_prio_battle_locations_before_1989]{ucdp_prio_battle_locations_before_1989()}}
The shares are based on reading Wikipedia and the documentation for the PRIO Battledeaths dataset, and amounts to a qualified impression of where most fighting occurred.
This is not an assessment of fatalities broken down in events. E.g., .5/.5 is used when fighting occurred on both sides of a contested border.
.1/.9 is used if most fighting happened on one side of the border, but with documented fighting on the other side. The code is made on a year
by year basis for each conflict_id, and the shares can change over years.
}
\keyword{datasets}
