% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\docType{data}
\name{cd}
\alias{cd}
\title{Creates a disk cache object that can be used with memoise}
\format{
An object of class \code{cache_disk} (inherits from \code{cachem}) of length 11.
}
\usage{
cd
}
\description{
Creates a disk cache object that can be used with memoise
}
\keyword{datasets}
