% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\name{check_ucdp_version}
\alias{check_ucdp_version}
\title{Test if a version number exists in the UCDP GED API}
\usage{
check_ucdp_version(version)
}
\arguments{
\item{version}{A character string}
}
\value{
A boolean
}
\description{
See https://ucdp.uu.se/apidocs/ for more information
}
\examples{
versions <- c("23.1", "26.2", "23.0.10")
versions <- sapply(versions, check_ucdp_version)
}
