% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{custom_gwcode_matches}
\alias{custom_gwcode_matches}
\title{Custom country to gwcode matches}
\format{
A named vector
}
\usage{
custom_gwcode_matches
}
\description{
A set of matches that fits well with `cshp_gw_modifications()` and `countrycode(custom_match = custom_gwcode_matches)`
Depending on whether you match from gwcode to something else, or from something to gwcode, you will need to swap
names and values of this vector
}
\keyword{datasets}
