% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\name{distribute_count}
\alias{distribute_count}
\title{Distributes a count over n periods as uniformly as possible}
\usage{
distribute_count(count, n)
}
\arguments{
\item{count}{A positive integer of the count that should be distributed}

\item{n}{Number of periods to distribute over}
}
\value{
A vector of distributed counts
}
\description{
Distributes from start to end, but ensures that there is at least one in the last observation if possible.
}
\examples{
distribute_count(3, 4)
distribute_count(NA, 3)
distribute_count(8, 3)
}
