% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdi.R
\name{download_wdi_uncached}
\alias{download_wdi_uncached}
\alias{download_wdi}
\title{Get data from World Development Indicators}
\usage{
download_wdi_uncached(
  indicators,
  start = 1960,
  end = lubridate::year(Sys.Date()),
  gwcode = TRUE,
  merge_fun = hablar::sum_
)

download_wdi(
  indicators,
  start = 1960,
  end = lubridate::year(Sys.Date()),
  gwcode = TRUE,
  merge_fun = hablar::sum_
)
}
\arguments{
\item{indicators}{Character vector of indicators codes. See the WDI::WDIsearch() function.
If you supply a named vector, the indicators will be automatically renamed:
‘c(’women_private_sector' = 'BI.PWK.PRVS.FE.ZS')'}

\item{start}{Start date in year integer format.}

\item{end}{End date in year integer format.}

\item{gwcode}{If true, use the gwcode identifier for countries}

\item{merge_fun}{If gwcode is true, some countries are merged. This sets the function used to merge data. See ?hablar::sum_ for default function.}
}
\value{
A data frame with WDI data
}
\description{
Get data from World Development Indicators
}
\section{Functions}{
\itemize{
\item \code{download_wdi()}: Get data from World Development Indicators, cached version

}}
\examples{
indicators <- c("wdi_pop" = "SP.POP.TOTL",
                "wdi_gdp_pp_con_us" = "NY.GDP.MKTP.PP.KD",
                "wdi_gdp_pp_cur_us" = "NY.GDP.MKTP.PP.CD")
df <- download_wdi(indicators)
}
\references{
http://arelbundock.com https://github.com/vincentarelbundock/WDI

See WDI indicator online for correct citation.
}
