% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshapes.R
\name{find_territorial_dependencies}
\alias{find_territorial_dependencies}
\title{Find territorial dependencies of a specific gwcode given a cShapes like dataset}
\usage{
find_territorial_dependencies(gwcode, gw)
}
\arguments{
\item{gwcode}{An integer or string with the gwcode of the country you want to find dependencies for}

\item{gw}{The cShapes data based on Gleditsch-Ward.}
}
\value{
An igraph graph
}
\description{
Find territorial dependencies of a specific gwcode given a cShapes like dataset
}
\examples{
gw <- cshp_gw_modifications()
system_which_includes_russia <- find_territorial_dependencies(365, gw)
plot(system_which_includes_russia)

}
