% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshapes.R
\name{get_custom_gwcode_match}
\alias{get_custom_gwcode_match}
\title{Helper function to get vector with custom gwcode matches}
\usage{
get_custom_gwcode_match(match_from_gwcode = TRUE)
}
\arguments{
\item{match_from_gwcode}{If true, returns a vector with gwcodes as names and country names as values, opposite if false}
}
\value{
A named vector
}
\description{
See `custom_gwcode_matches`
}
\examples{
get_custom_gwcode_match(match_from_gwcode = TRUE)
get_custom_gwcode_match(match_from_gwcode = FALSE)
}
