% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prio_battledeaths.R
\name{get_prio_brd}
\alias{get_prio_brd}
\title{Downloads PRIO Battle Deaths Dataset 1946-2008 version 3}
\usage{
get_prio_brd()
}
\value{
A data frame with the Battle Deaths data.
}
\description{
Please cite:
Bethany Lacina and Nils Petter Gleditsch, 2005. "Monitoring Trends in Global Combat: A New Dataset of Battle Deaths." European Journal of
Population: 21(2–3): 145–166. The data are available at http://www.prio.no/CSCW/Datasets/Armed-Conflict/Battle-Deaths/
}
\details{
Updated UCDP/PRIO conflict id's are merged in for convenience.
}
\examples{
brds <- get_prio_brd()
}
