% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwt.R
\name{get_pwt_uncached}
\alias{get_pwt_uncached}
\alias{get_pwt}
\title{Get Penn World Table data}
\usage{
get_pwt_uncached(pwt_version = "10.01", gwcode = TRUE)

get_pwt(pwt_version = "10.01", gwcode = TRUE)
}
\arguments{
\item{pwt_version}{A string with the version of PWT you want to downlaod.}
}
\value{
A data frame with PWT data
}
\description{
Get Penn World Table data
}
\section{Functions}{
\itemize{
\item \code{get_pwt()}: Get Penn World Table data, cached version

}}
\examples{
pwt <- get_pwt()
}
\references{
Feenstra, Robert C., Robert Inklaar and Marcel P. Timmer (2015), "The Next Generation of the Penn World Table" American Economic Review, 105(10), 3150-3182, available for download at www.ggdc.net/pwt
}
