% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\name{get_ucdp_ged}
\alias{get_ucdp_ged}
\title{Get the UCDP GED data from their API}
\usage{
get_ucdp_ged(version)
}
\arguments{
\item{version}{A character string denoting the version of UCDP GED you want.}
}
\value{
A data frame with UCDP GED data.
}
\description{
See https://ucdp.uu.se/apidocs/ for more information
}
\details{
The function cache the results using memoise::memoise.
}
\examples{
stable_version <- latest_ucdp_version(type = "stable")
ucdp <- get_ucdp_ged(stable_version)
}
