% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\name{get_ucdp_uncached}
\alias{get_ucdp_uncached}
\alias{get_ucdp}
\title{Get the UCDP data from their API}
\usage{
get_ucdp_uncached(version, dataset = "gedevents")

get_ucdp(version, dataset = "gedevents")
}
\arguments{
\item{version}{A character string denoting the version of UCDP GED you want.}

\item{dataset}{Which dataset to search. One of c("gedevents", "ucdpprioconflict", "dyadic", "nonstate", "onesided", "battledeaths")}
}
\value{
A data frame with UCDP GED data.
}
\description{
See https://ucdp.uu.se/apidocs/ for more information
}
\details{
The function cache the results using memoise::memoise.
}
\section{Functions}{
\itemize{
\item \code{get_ucdp()}: Get the UCDP GED data from their API

}}
\examples{
stable_version <- latest_ucdp_version(type = "stable")
ged <- get_ucdp(stable_version)
}
