% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem.R
\name{get_vdem}
\alias{get_vdem}
\title{Helper function to get V-Dem variables}
\source{
https://www.v-dem.net/
}
\usage{
get_vdem(..., .fun = mean, na.rm = TRUE, impute = TRUE, merge_countries = TRUE)
}
\arguments{
\item{...}{<\code{\link[=dplyr::dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables. See `vdemdata::codebook$clean_tag` for valid expressions.}

\item{.fun}{The function used to summarize observations for mergers.}

\item{impute}{If true the following things are added:
Includes data from the larger state 'vdem_id' in the from-to period for all states with include_id as coded in poldat::vdem_notes.
Uses USA coding for Puerto Rico.
Uses France coding for New Caledonia and French Polynesia.
Belize, Bahamas, Western Sahara, and Brunei are kept missing.}

\item{merge_countries}{If true, the following countries are merged (with equal weights using he .fun):
West Bank and Gaza from 1967 onwards into 699.
Czechia and Slovakia (Czechoslovakia) from 1946 to 1992.
Bosnia and Herzegovina, Kosovo, North Macedonia, Montenegro, Serbia, and Croatia (Yugoslavia) from 1946-1990
Bosnia and Herzegovina, Kosovo, Montenegro, Serbia, Croatia (Yugoslavia) in 1991
Kosovo, Montenegro, Serbia (Yugoslavia) from 1992-2005
Serbia and Kosovo (Serbia) from 2006-2007
Subsequently, only observations that fit the gw_panel template with gw <- cshp_gw_modifications() are included
E.g., instead of having both North and South Vietnam from 1946-present, this will remove South Vietnam after 1974.
Merging countries like this makes sure that the double observations from the impute step are removed again,
and that the country structure is identical to the modified cShapes.}
}
\description{
Use `vdemdata::vdem`, but matches (as best as possible) to the cShapes system used in poldat.
The system used by V-Dem is not static, but does not match perfectly up with poldat.
For example, West Bank is not coded from 1950-1966 in V-Dem because of the Jordanian
occupation, and treats West Bank as both West Bank and Gaza from 1994-2007. cShapes
codes West Bank and Gaza independently until 1967, and then as part of Israel since,
whilst there is an option in poldat to treat West Bank and Gaza as a common unit after
1967. This function merge states that are treated as a single unit in poldat, but multiple in
V-Dem. It likely misses many issues, but tries to accommodate at least political units currently
present (as coded by poldat/cShapes).
}
\details{
Current states with > 250 000 population that are missing in V-Dem:
 Puerto Rico, Bahamas, Belize, Western Sahara, Brunei, New Caledonia and Dependencies, French Polynesia

Filter data before 1946 for the moment (as there are likely many more inconsistencies before then).

More on the imputation idea: When a territory is included in another state-coding, there is no data for that territory separately in V-Dem.
}
\examples{
df <- get_vdem(v2x_libdem)
df <- get_vdem(v2x_libdem, v2x_regime, .fun = min)

}
\references{
Lührmann, Anna, Nils Düpont, Masaaki Higashijima, Yaman Berker Kavasoglu, Kyle L. Marquardt, Michael Bernhard, Holger Döring, Allen Hicken, Melis Laebens, Staffan I. Lindberg, Juraj Medzihorsky, Anja Neundorf, Ora John Reuter, Saskia Ruth-Lovell, Keith R. Weghorst, Nina Wiesehomeier, Joseph Wright, Nazifa Alizada, Paul Bederke, Lisa Gastaldi, Sandra Grahn, Garry Hindle, Nina Ilchenko, Johannes von Römer, Steven Wilson, Daniel Pemstein, Brigitte Seim. 2020. Varieties of Party Identity and Organization (V-Party) Dataset V1. Varieties of Democracy (V-Dem) Project. https://doi.org/10.23696/vpartydsv1
}
