% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\name{gw_ged_uncached}
\alias{gw_ged_uncached}
\alias{gw_ged}
\title{Complete country-time panel data with UCDP GED and modified cShapes}
\usage{
gw_ged_uncached(static_date = NULL, ...)

gw_ged(static_date = NULL, ...)
}
\arguments{
\item{static_date}{If set, then pretend the world borders always looked like cShapes at static_date.}

\item{...}{
  Arguments passed on to \code{\link[=cshp_gw_modifications]{cshp_gw_modifications}}, \code{\link[=ucdp_ged_panel]{ucdp_ged_panel}}
  \describe{
    \item{\code{western_sahara}}{Morocco is coded without Spanish Sahara, and Spanish Sahara becomes Western Sahara after 13 Nov 1975.
cShapes codes the whole of Spanish Sahara as becoming a part of Morocco from Nov 1975.
Although Morocco did occupy parts of Spanish Sahara in 1975, the fight against the Polisario
Front (Western Sahara) ended in a cease-fire in 1991 and territorial claims have never
been resolved. Fighting resumed in 2020. Morocco now controls most of Western Sahara,
and importantly the costal areas ("Southern Provinces").}
    \item{\code{morocco_protectorate}}{Recodes Morocco from "N/A" to "protectorate" from 1904-01-02.
The only use of "N/A" in cshapes is Morocco from 1904-1912. The Entente Cordiale of 1904,
the Algeciras Conference in 1906, the Agadir incidence in 1911, and the Treaty of Fez in 1912
pulls Morocco into France's sphere of influence. Although not formally a protectorate
until the Treaty of Fez, it is not unreasonable to code this as a protectorate instead
of "N/A" from 1904.}
    \item{\code{palestine}}{Israel is not granted Gaza, West Bank or Sinai due to length of occupation.
After 6-day war, Israel becomes occupant ("owner") of West Bank and Gaza, instead of gaining territories as coded in cShapes.
Gleditsch-Ward has a 10-year occupation rule saying that any territory occupied more than 10 years falls
to the occupier. In cShapes, this means that Gaza, West Bank, and Sinai becomes part of Israel
from 1967 until 1979 (when Egypt formally got Sinai back, although troops stayed until Jan 1982).
Palestine is given the GW code 699.}
    \item{\code{soviet_25dec}}{Moves the dissolution of Soviet from 21 Dec 1991 to 25 Dec 1991.
A number of dates could be used to denote the dissolution of Soviet. 21 December was
the date of the Alma-Ata protocol. 25 Dec was the date of Gorbachev's speech, whilst
26 Dec was when Soviet recognized Ukraine, Armenia, Azerbaijan, and Kazakhstan. The issue
with using 21 Dec as the date of Soviet dissolution is that cShapes codes the start of the
new states from 26 Dec. So, there are five days where Ukraine, Armenia, Azerbaijan, and Kazakhstan
territories have no state control according to cShapes. In certain applications, having full
temporal contiguity is very useful. So this is a practical code change, more than anything.
(E.g., Ukraine celebrates their independence day on 24 August, not 26 December.)}
    \item{\code{yugoslavia_4jun}}{Moves the dissolution of Yugoslavia to 4 June 2006, the start of Montenegro and Serbia to 5 June.
Montenegro declared independence on 3 June, whilst Serbia declared independence 5 June.
cShapes 2.0 codes the end of Yugoslavia 2006-06-02, the start of Serbia 2006-06-06
and the start of Montenegro 2006-06-12. GW codes the end of Yugoslavia 2006-06-04, the start of Serbia 2006-06-05,
and the start of Montenegro 2006-06-03. Using the dissolution of the larger unit is consistent with how Soviet is treated.}
    \item{\code{kosovo_17feb}}{Moves the start date of Kosovo to 2008-02-17. This is probably just a coding error (in cshapes 2.0
it is 2008-02-20). Kosovo unilaterally declared independence 17 Feb, and that is also the start date in GW.}
    \item{\code{senegal_22sep}}{Senegal withdrew from the Mali Federation 20 September 1960, not 21 June 1960. Could possibly have
confused the date with when the Mali Federation became independent from France (20 June 1960). The Declaration of the
Republic of Mali happened 22 September 1960, which is consistent with start dates of independent Mali in both cShapes
and GW. To be consistent with other coding (e.g., Soviet Union), move start of Senegal from 20 September to 22 September.
The geometry used to represent Mali Federation from June 1960 to September 1960 (fid = 236) appears to be wrong, and is
swapped with fid = 235.}
    \item{\code{kuwait_protectorate}}{Before independence, Kuwait was a British Protectorate. This is not present in cShapes 2.0. This
adds this. Uqair Protocol of 1922 defined borders between Mandatory Iraq, Nejd, and Kuwait. Before then, the borders where
disputed, with the Kuwait-Najd War breaking out after the Ottoman Empire was defeated in WW1 and the Anglo-Ottoman Convention of
1913 was invalidated. Kuwait existed as a British protectorate from 23 January 1899, but this was disputed by the Ottomans. Note
that this still leaves a stretch between 1914 - 1922 without any coding of this territory. But then, the Anglo-Ottoman Convention
of 1913 does not make it easy for coders.}
    \item{\code{bahrain_protectorate}}{Before independence, Bahrain was a British Protectorate. This is not present in cShapes 2.0. This
adds this. On 22 December 1880, Britain signed a treaty with Bahrain that could be argued to be the start of a protectorate. A
treaty in 1892 further limited Bahrain's power, and in the Anglo-Ottoman convention of 1913, the Ottomans renounced all claims
they had to Bahrain. Use 22 December 1880 as the start date for the British Protectorate. Earlier dates could be
possible start dates, e.g., May 1861 (first treaty with Britain) and 2 December 1869 (start of Isa Bin Ali Al Khalifa's reign).}
    \item{\code{version}}{String denoting the UCDP GED version to use.}
    \item{\code{time_interval}}{String denoting the temporal interval to use.}
    \item{\code{state_based}}{Boolean if state_based events should be included}
    \item{\code{non_state}}{Boolean if non_state events should be included}
    \item{\code{one_sided}}{Boolean if one_sided events should be included}
    \item{\code{drop_poor_precision}}{Boolean if events with temporal precision poorer than time_interval should be included}
    \item{\code{test}}{Boolean to run code with smaller sample, used for testing only}
  }}
}
\value{
Data frame with country information, and average date precision (of UCDP GED events) and sum of battle-deaths.
}
\description{
This will override the country_id used by UCDP, and instead use the gwcode in cShapes.
For events that fall outside of cShapes polygons, this function falls back to the UCDP gazetteer coding.
The benefit of using cShapes is that this makes it possible to see events happening in e.g., Western Sahara or Palestine.
It also makes the country-code consistent across various datasets that is based on cShapes. While UCDP country_id
and cShapes mostly align, there are a few cases where they do not. E.g., UCDP codes events in Namibia in 1989 when it was a
South African colony with the gwcode of South Africa, whilst cShapes uses the Namibia code.
}
\section{Functions}{
\itemize{
\item \code{gw_ged()}: Get the UCDP GED data from their API

}}
\examples{
ged <- gw_ged(western_sahara = TRUE,
               morocco_protectorate = TRUE,
               palestine = TRUE,
               soviet_25dec = TRUE,
               version = "23.1",
               time_interval = "year",
               state_based = TRUE,
               non_state = TRUE,
               one_sided = TRUE,
               drop_poor_precision = TRUE,
               test = TRUE)
}
