% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\name{latest_ucdp_version}
\alias{latest_ucdp_version}
\title{Find the latest version of the UCDP data}
\usage{
latest_ucdp_version(dataset = "gedevents", type = "stable")
}
\arguments{
\item{dataset}{Which dataset to search. One of c("gedevents", "ucdpprioconflict", "dyadic", "nonstate", "onesided", "battledeaths")}

\item{type}{Either 'stable' or 'candidate' (only relevant for "gedevents" dataset)}
}
\value{
A character string representing the latest version of UCDP data.
}
\description{
See https://ucdp.uu.se/apidocs/ for more information
}
\examples{
stable_version <- latest_ucdp_version(type = "stable")
ged <- get_ucdp(stable_version)
}
