% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_world_duckdb.R
\name{poldat_db_path}
\alias{poldat_db_path}
\title{Default path of the poldat DuckDB research database}
\usage{
poldat_db_path()
}
\value{
A file path inside \code{rappdirs::user_data_dir("R-poldat")}.
}
\description{
Default path of the poldat DuckDB research database
}
