% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{static_world}
\alias{static_world}
\title{Panel data with static country ids}
\format{
A data frame with 13783 rows and 7 columns:
\describe{
  \item{gwcode}{Slightly modified cShapes 2.0 gwcodes}
  \item{year}{The year of the events}
  \item{best}{Best estimate of battle-deaths}
  \item{low}{Low estimate of battle-deaths}
  \item{high}{High estimate of battle-deaths}
  \item{v2x_libdem}{Liberal democracy score}
  \item{v2x_regime}{Regimes of the World class}
}
}
\source{
[PRIO Battle-deaths 3.1](https://www.prio.org/data/1)

[UCDP GED 23.1](https://ucdp.uu.se)

[cShapes 2.0](https://icr.ethz.ch/data/cshapes/)

[Varieties of Democracy 14](https://v-dem.net/)
}
\usage{
static_world
}
\description{
Uses `area_weighted_synthetic_data(static_year = 2019)` to create panel data with static country ids.
V-Dem data using `get_vdem()` and UCDP battle-related deaths from `ucdpbrds`. Please see these
functions for more details and references.
}
\keyword{datasets}
