% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cshapes.R
\name{territorial_dependencies_base}
\alias{territorial_dependencies_base}
\alias{territorial_dependencies}
\title{Creates a network graph of territorial dependencies from the cShapes Gleditsch-Ward data}
\usage{
territorial_dependencies_base(gw)

territorial_dependencies(gw)
}
\arguments{
\item{gw}{The cShapes data based on Gleditsch-Ward.
Using cshp_gw_modifications() with soviet_25dec = TRUE is recommended.}
}
\value{
An igraph network graph
}
\description{
Any country that contains or overlaps with another country in space, and is adjacent (1 day) in time
are defined as neighbors. From this list of neighbors, a network graph is built.
}
\section{Functions}{
\itemize{
\item \code{territorial_dependencies()}: Creates a network graph of territorial dependencies from the cShapes Gleditsch-Ward data

}}
\examples{
gw <- cshp_gw_modifications(soviet_25dec = TRUE)
g <- territorial_dependencies(gw)

}
