% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\name{ucdp_ged_panel}
\alias{ucdp_ged_panel}
\title{Distributes/aggregates UCDP GED at the country-level over a chosen time-interval}
\usage{
ucdp_ged_panel(
  version,
  time_interval,
  state_based,
  non_state,
  one_sided,
  drop_poor_precision,
  test = FALSE,
  ...
)
}
\arguments{
\item{version}{String denoting the UCDP GED version to use.}

\item{time_interval}{String denoting the temporal interval to use.}

\item{state_based}{Boolean if state_based events should be included}

\item{non_state}{Boolean if non_state events should be included}

\item{one_sided}{Boolean if one_sided events should be included}

\item{drop_poor_precision}{Boolean if events with temporal precision poorer than time_interval should be included}

\item{test}{Boolean to run code with smaller sample, used for testing only}

\item{...}{Additional parameters}
}
\value{
Data frame with longitude, latitude, country_id, average date precision and sum of battle-deaths.
}
\description{
Distributes/aggregates UCDP GED at the country-level over a chosen time-interval
}
\examples{
ged <- ucdp_ged_panel(version = "23.1",
                      time_interval = "year",
                      state_based = TRUE,
                      non_state = TRUE,
                      one_sided = TRUE,
                      drop_poor_precision = TRUE,
                      test = TRUE)
}
\seealso{
\code{\link[=seq.Date]{seq.Date()}} for valid time_interval strings.
}
