% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp_prio.R
\name{ucdp_long_cy_panel}
\alias{ucdp_long_cy_panel}
\title{A country-year panel data with battle-locations and intensity from 1946-present}
\usage{
ucdp_long_cy_panel(version = "23.1")
}
\value{
A data frame with country-year panel data with battle-locations and intensity from 1946-present
}
\description{
Is based on cShapes 2.0, UCDP/PRIO ACD (1946-88) and UCDP GED (1989-present).
}
\details{
Using modifications to cshapes using `cshp_gw_modifications()`.
Using re-coded battle-locations data from `ucdp_prio_battle_locations_before_1989()`

Use UCDP GED point locations to determine country battle location based on the modified cShapes data
(instead of UCDP GED's own country_id code).

Please cite:

Gleditsch, Nils Petter; Peter Wallensteen, Mikael Eriksson, Margareta Sollenberg & Håvard Strand
(2002) Armed Conflict 1946–2001: A New Dataset. Journal of Peace Research 39(5): 615–637.

Sundberg, Ralph, and Erik Melander, 2013, “Introducing the UCDP Georeferenced
Event Dataset”, Journal of Peace Research, vol.50, no.4, 523-532

Schvitz, Guy, Seraina Rüegger, Luc Girardin, Lars-Erik Cederman, Nils Weidmann, and Kristian Skrede Gleditsch. 2022.
“Mapping The International System, 1886-2017: The CShapes 2.0 Dataset.” Journal of Conflict Resolution 66(1): 144–61.
}
\examples{
df <- ucdp_long_cy_panel()
}
