% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp_prio.R
\name{ucdp_prio_battle_locations_before_1989}
\alias{ucdp_prio_battle_locations_before_1989}
\title{Codes the gwcode of where majority of battles occurred in a conflict-dyad in the UCDP/PRIO ACD before 1989}
\usage{
ucdp_prio_battle_locations_before_1989()
}
\value{
A data frame with UCDP/PRIO ACD data version 23.1 plus variable "battle_loc".
}
\description{
gwcodes are based on cShapes 2.0 (with modifications in `cshp_gw_modifications'). This coding is
using UCDP/PRIO ACD version 23.1. Supplement with data from UCDP GED after 1989.
}
\details{
Coded by Jonas Vestby, 22.01.2024. All errors are mine.
}
\examples{
ucdp_prio <- ucdp_prio_battle_locations_before_1989()
}
