% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wcde_gwcode}
\alias{wcde_gwcode}
\title{Historical population and education data from Wittgenstein Centre Data Explorer v.2}
\format{
A data frame with 2 694 rows and 12 columns:
\describe{
 \item{gwcode}{gwcode based on the modified CShapes 2.0, see poldat::cshp_gw_modifications}
 \item{year}{The year in four digits}
 \item{secprop}{The proportion of the population with completed upper secondary education}
 \item{priprop}{The proportion of the population with completed primary education}
 \item{psecprop}{The proportion of the population with completed post-secondary education}
 \item{tdr}{Total dependency ratio: (youth + elderly) / working.}
 \item{ydr}{Youth dependency ratio: youth / working}
 \item{odr}{Old-age dependency ratio: elderly / working}
 \item{youth}{Population between 0-14}
 \item{working}{Population between 15-64}
 \item{elderly}{Population from 65+}
 \item{tot_pop}{Total population}
}
}
\usage{
wcde_gwcode
}
\description{
Based on wcde::past_epop, but sets country unit as gwcode, and estimates a set of variables based on this.
Deviation can arise from wcde as some countries are merged when using gwcode. Data every 5th year.
}
\references{
Lutz, Wolfgang, Anne Goujon, Samir Kc, Marcin Stonawski, and Nikolaos Stilianakis. Demographic and human capital scenarios for the 21st century: 2018 assessment for 201 countries. Publications Office of the European Union, 2018.
}
\keyword{datasets}
