% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem.R
\name{get_vdem}
\alias{get_vdem}
\title{Helper function to get V-Dem variables}
\source{
https://www.v-dem.net/
}
\usage{
get_vdem(..., .fun = mean, na.rm = TRUE)
}
\arguments{
\item{...}{<\code{\link[=dplyr::dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables. See `vdemdata::codebook$clean_tag` for valid expressions.}

\item{.fun}{The function used to summarize observations for mergers.}
}
\description{
Use `vdemdata::vdem`, but matches (as best as possible) to the cShapes system used in poldat.
The system used by V-Dem is not static, but does not match perfectly up with poldat.
For example, West Bank is not coded from 1950-1966 in V-Dem because of the Jordanian
occupation, and treats West Bank as both West Bank and Gaza from 1994-2007. cShapes
codes West Bank and Gaza independently until 1967, and then as part of Israel since,
whilst there is an option in poldat to treat West Bank and Gaza as a common unit after
1967. This function merge states that are treated as a single unit in poldat, but multiple in
V-Dem. It likely misses many issues, but tries to accommodate at least political units currently
present (as coded by poldat/cShapes).
}
\details{
Filter data before 1946 for the moment.

Uses Jordan coding for West Bank between 1950-67.
Uses USA coding for Puerto Rico (as that is completely missing from V-Dem at the moment.)
}
\examples{
df <- get_vdem(v2x_libdem)
df <- get_vdem(v2x_libdem, v2x_regime, .fun = min)

}
\references{
Lührmann, Anna, Nils Düpont, Masaaki Higashijima, Yaman Berker Kavasoglu, Kyle L. Marquardt, Michael Bernhard, Holger Döring, Allen Hicken, Melis Laebens, Staffan I. Lindberg, Juraj Medzihorsky, Anja Neundorf, Ora John Reuter, Saskia Ruth-Lovell, Keith R. Weghorst, Nina Wiesehomeier, Joseph Wright, Nazifa Alizada, Paul Bederke, Lisa Gastaldi, Sandra Grahn, Garry Hindle, Nina Ilchenko, Johannes von Römer, Steven Wilson, Daniel Pemstein, Brigitte Seim. 2020. Varieties of Party Identity and Organization (V-Party) Dataset V1. Varieties of Democracy (V-Dem) Project. https://doi.org/10.23696/vpartydsv1
}
