% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp.R
\name{latest_ucdp_version}
\alias{latest_ucdp_version}
\title{Find the latest version of the UCDP GED data}
\usage{
latest_ucdp_version(type = "stable")
}
\arguments{
\item{type}{Either 'stable' or 'candidate'}
}
\value{
A character string representing the latest version of UCDP data.
}
\description{
See https://ucdp.uu.se/apidocs/ for more information
}
\examples{
stable_version <- latest_ucdp_version(type = "stable")
ucdp <- get_ucdp_ged(stable_version)
}
